requirements.txt
absl-py==2.4.0
altair==6.0.0
astunparse==1.6.3
attrs==25.4.0
black==26.1.0
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
imbalanced-learn==0.14.1
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kaggle==1.8.3
kagglesdk==0.1.15
keras==3.13.2
kiwisolver==1.4.9
libclang==18.1.1
librt==0.7.8
llvmlite==0.46.0
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mypy==1.19.1
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.16.0
numba==0.63.1
numpy==2.3.5
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pandas==2.3.3
pathspec==1.0.4
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.2
protobuf==6.33.5
pyarrow==23.0.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytokens==0.4.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rich==14.3.2
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
shap==0.50.0
six==1.17.0
sklearn-compat==0.1.5
slicer==0.0.8
smmap==5.0.2
streamlit==1.54.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
types-requests==2.32.4.20260107
types-tqdm==4.67.3.20260205
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
webencodings==0.5.1
Werkzeug==3.1.5
wrapt==2.1.1
